Checking for wrong inputs, and various ways of executing entities

+++ F4Server.c
@@ -11,6 +11,7 @@ int * shm_info_attach;
 int count_sigint = 0;
 char symbol[2] = {MATRIX_DEFAULT_CHAR, MATRIX_DEFAULT_CHAR};
 
+
 void delete_all(){
     shmctl(shm_matrix_id, IPC_RMID, NULL);
     shmctl(shm_info, IPC_RMID, NULL);
@@ -22,23 +23,28 @@ void delete_all(){
 /// @param string The string to put in perror
 void perror_exit_server(char * string){
     perror(string);
-    delete_all();
+    //delete_all();
     exit(-1);
 }
 
 /// @brief Handler for the SIGUSR1
 /// @param sig The value of the signal
 void sigusr1_handler(int sig){
-    if (shm_info_attach[9] == symbol[0]){  
-        printf("%c is out of the game!\n", symbol[0]);  
-        printf("%c is the WINNER!", symbol[1]);  
-        kill(shm_info_attach[5], SIGUSR2);
-    }else{
-        printf("%c is out of the game!\n", symbol[1]);  
-        printf("%c is the WINNER!", symbol[0]); 
-        kill(shm_info_attach[4], SIGUSR2); 
+    if (shm_info_attach[0]==2){
+        if (shm_info_attach[9] == symbol[0]){  
+            printf("%c is out of the game!\n", symbol[0]);  
+            printf("%c is the WINNER!", symbol[1]);  
+            kill(shm_info_attach[5], SIGUSR2);
+        }
+        else{
+            printf("%c is out of the game!\n", symbol[1]);  
+            printf("%c is the WINNER!", symbol[0]); 
+            kill(shm_info_attach[4], SIGUSR2); 
+        }
     }
-    delete_all();
+    else{
+        printf("The player closed the game, please restart the server...\n");
+    }
     exit(0);
 }
 
@@ -55,13 +61,17 @@ void sigint_handler(int sig){
             }
             
         }
-        delete_all();
         exit(-1);
     }else if (count_sigint == 1){ /* Warning. */
         printf("Next CTRL+C will kill the game!\n");
     }
 }
 
+/// @brief Initializes the info matrix
+/// @param shm_info_attach the pointer to the matrix
+/// @param N the number of rows
+/// @param M the number of columns
+/// @param timer the value of the timer
 void shm_info_init(int * shm_info_attach, int N, int M, int timer){
     shm_info_attach[0]  = 0;
     shm_info_attach[1]  = symbol[0];
@@ -77,6 +87,18 @@ void shm_info_init(int * shm_info_attach, int N, int M, int timer){
     shm_info_attach[11] = 0;
 }
 
+/// @brief Checks if a string is Numeric
+/// @param string the string to check
+/// @return 1 if numeric, 0 otherwise
+int isNumeric(char* string){
+    int count = 0;
+    while (string[count]!='\0'){
+        if (string[count] <'0' || string[count] > '9')return 0;
+        count++;
+    }
+    return 1;
+}
+
 
 
 int main(int argc, char *argv[])
@@ -87,17 +109,22 @@ int main(int argc, char *argv[])
         printf("The corretct sinxtax should be: ./F4Server n_rows n_colums symbol1 symbol2 \n");
         return -1;
     }
-    print_banner();
 
-    /* Getting the timer for the clients, 0 no timer. */
-    int timer;
-    printf("Insert the number of seconds for each turn (0 for no timer): ");
-    scanf("%d", &timer);
     
+
     /* Getting the info from the command line and checking if they are acceptable. */
+    if(!isNumeric(argv[1])){
+        printf("The number of Rows is not a number...\n");
+        exit(-1);
+    }
     const int N = string_to_int(argv[1]);
     if (N < 5){
         printf("Number of Rows not acceptable...\n");
+        exit(-1);
+    }
+
+    if(!isNumeric(argv[2])){
+        printf("The number of Columns is not a number...\n");
         exit(-1);
     }
     const int M = string_to_int(argv[2]);
@@ -105,9 +132,25 @@ int main(int argc, char *argv[])
         printf("Number of Columns not acceptable...\n");
         exit(-1);
     }
+    
+    /* Checking if the symbols are valid. */
+    if((strlen(argv[3])==1 && strlen(argv[4])==1) && (argv[3][0]!=argv[4][0])){
+        symbol[0] = argv[3][0];
+        symbol[1] = argv[4][0];
+    }
+    else{
+        printf("The symbols must be 1 char each and different!\n");
+        exit(-1);
+    }
 
-    symbol[0] = argv[3][0];
-    symbol[1] = argv[4][0];
+    /* Forza4 Logo. */
+    print_banner();
+
+    /* Setting the exit function to delete all the ipcs */
+    if (atexit(delete_all) == -1){
+        perror("Error setting delete_all up...");
+        exit(-1);
+    }
 
     /* Handling all the signals. */
     if (signal(SIGINT, sigint_handler) == SIG_ERR){
@@ -119,6 +162,30 @@ int main(int argc, char *argv[])
         perror("Error Handling SIGUSR1!\n");
         exit(-1);
     }
+
+    /* Setting the timer for the turns. */
+    int timer;
+    int checkValidTimer;
+    do{
+        checkValidTimer = 1;
+        printf("Insert the number of seconds for each turn (0 for no timer): ");
+
+
+        if(!scanf(" %d", &timer)){
+            int c;
+            while ((c = getchar()) != '\n' && c != EOF) {}
+            checkValidTimer = 0;
+            printf("The timer should be a number...\n");
+        }
+
+        if (timer<0){
+            printf("The timer cannot be negative...\n");
+            checkValidTimer = 0;
+        }
+
+    }while(!checkValidTimer);
+
+    
     
     /* Sops for the sync semaphores. */
     struct sembuf sops[3];
@@ -144,16 +211,20 @@ int main(int argc, char *argv[])
     if ((matrix_pointer = (int *) shmat(shm_matrix_id, NULL, 0)) == NULL){
         perror_exit_server("Matrix Shared Memory Attach...");
     }
+
+    int shm_info_key = ftok("key.txt", SHMINFO_KEY);
+
+    /* Creating the shm that store various information. */
+    if ((shm_info = shmget(shm_info_key, sizeof(int) * 12, IPC_CREAT | 0777 | IPC_EXCL)) == -1){
+        perror_exit_server("A server already exists...");
+    }
     
     /* Initializing the matrix. */
     matrix_init(matrix_pointer, N, M);
 
     printf("Initializing matrix...\n");
 
-    /* Creating the shm that store various information. */
-    if ((shm_info = shmget(SHMINFO_KEY, sizeof(int) * 12, IPC_CREAT | 0777 | IPC_EXCL)) == -1){
-        perror_exit_server("Info Shared Memory...");
-    }
+    
 
     /* Attaching to the previous shm. */
     if ((shm_info_attach = (int *) shmat(shm_info, NULL, 0)) == NULL){
@@ -165,8 +236,10 @@ int main(int argc, char *argv[])
 
     printf("Initializing infos...\n");
 
+    int sem_sync_key = ftok("key.txt", SEMSYNC_KEY);
+
     /* Creating the sem sync semaphore. */
-    if ((sem_sync = semget(SEMSYNC_KEY, 3, IPC_CREAT | IPC_EXCL | 0777))== -1){
+    if ((sem_sync = semget(sem_sync_key, 3, IPC_CREAT | IPC_EXCL | 0777))== -1){
         perror_exit_server("Semaphore Sync...");
     }
 
@@ -182,8 +255,10 @@ int main(int argc, char *argv[])
 
     printf("Setting semsync...\n");
 
+    int sem_mutex_key = ftok("key.txt", SEMMUTEX_KEY);
+
     /* Setting the mutex semaphore. */
-    if ((sem_mutex = semget(SEMMUTEX_KEY, 1, IPC_CREAT | IPC_EXCL | 0777))== -1)
+    if ((sem_mutex = semget(sem_mutex_key, 1, IPC_CREAT | IPC_EXCL | 0777))== -1)
         perror_exit_server("Semaphore Mutex...");
 
     arg.val = 1;
@@ -195,7 +270,11 @@ int main(int argc, char *argv[])
     printf("Waiting for players...\n");
 
     if ((semop(sem_sync, &sops[0], 1)) == -1){
-        perror_exit_server("Locking Server...");
+        if (errno == EINTR){
+                    if (semop(sem_sync, &sops[0], 1) == -1)
+                        perror_exit_server("Error locking myself...");
+        }else perror_exit_server("Error locking myself...");
+                    
     }
     
     /* Declarations. */
@@ -275,6 +354,13 @@ int main(int argc, char *argv[])
 
     printf("Game Over!\n");
     
-    delete_all();
+    //delete_all();
     return 0;
 }
+
+
+/************************************
+*VR471346, VR471414, VR471404
+*Aldegheri Alessandro, Venturi Davide, Zerman Nicolò
+*15/04/2023
+*************************************/
